refactor(cache): centralize dSpark tap frontier
